RaspberryPi_Files/requirements.txt
scikit-learn
kafka-python
joblib
requests
toml